printing sendto return

+++ rdp_server.c
@@ -158,7 +158,7 @@ int sendSYN()
 	generateHeaderSYN(header);
 	printf("%s\n", header);
 
-	ssize_t s = sendto(sock, header, strlen(header), 0, (struct sockaddr*)&sa, sizeof sa);
+	int s = sendto(sock, header, strlen(header), 0, (struct sockaddr*)&sa, sizeof sa);
 
 	printf("sendto return: %d\n", s);
 	syn_packs_sent += 1;
